# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a customer profile api. ",
        "version": "1.0.0",
        "title": "Qantas Customer Profile API",
        "contact": {
            "email": "ng.quentin@gmail.com"
        }
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/profile": {
            "post": {
                "tags": [
                    "profile"
                ],
                "summary": "Add a new profile",
                "description": "Used to create a new profile",
                "operationId": "addProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Profile object that needs to be sent to backend CRM",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Profile"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/profile/{profileId}": {
            "parameters": [
                {
                    "name": "profileId",
                    "in": "path",
                    "description": "ID of profile of interest",
                    "required": true,
                    "type": "integer",
                    "format": "int64"
                }
            ],
            "put": {
                "tags": [
                    "profile"
                ],
                "summary": "Update an existing profile",
                "description": "Update profile",
                "operationId": "updateProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Profile object that needs sent to CRM",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Profile"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Profile not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "profile"
                ],
                "summary": "Deletes a profile",
                "description": "",
                "operationId": "deleteProfile",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Address": {
            "type": "object",
            "required": [
                "addressline1",
                "postcode",
                "suburb"
            ],
            "properties": {
                "addressline1": {
                    "type": "string",
                    "example": "5.01 Level 5 Arch street"
                },
                "addressline2": {
                    "type": "string"
                },
                "postcode": {
                    "type": "string",
                    "maxLength": 4,
                    "minLength": 4
                },
                "suburb": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "home",
                        "work"
                    ]
                },
                "email": {
                    "type": "string",
                    "format": "email"
                }
            }
        },
        "Profile": {
            "type": "object",
            "required": [
                "firstname",
                "lastname"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "firstname": {
                    "type": "string",
                    "example": "Bruce"
                },
                "lastname": {
                    "type": "string",
                    "example": "Wayne"
                },
                "dateofbirth": {
                    "type": "string",
                    "format": "date"
                },
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Address"
                    }
                }
            }
        }
    }
}